Clean up alignment, reframe branch2 implementation, add eval metrics, add continue training logic, and helpers for debugging
1. Clean up redundant codes in alignment.py to make sure code brevity
2. We notice that alignment module will yield different result when inferencing one sample and inferencing a batch of samples, in which
the first is correct. However, during training time we have to use batch inferencing and results in wrong prediction on alignment embeddings.
The wordaround is to produce subtraction directly from the video alignment module and save it to the pickle file, then directly load it in dataloader.
Will clean up of video aligning codes after.
3. Add evaluation metrics (b1,b4,rouge,cider,bertscore) to be the same as what motionGPT have done. Becuase using nlpmetricverse will cause a cycle
import error in evaluate.py. Rename it to evaluation.py.
4. Some helper codes to help debugging (simply change the import path in net/**)
5. Visualization codes added in utils/dtw and utils/visualize

Signed-off-by: Tommy <c1l1mo600@gmail.com>