Add the difference setting Padding
If the `DIFFERENCE_SETTING` in `get_difference_feature()` is
`Subtraction`, then we get the difference between user embedding and
standard embedding. If the `DIFFERENCE_SETTING` is `Padding`, then we
get the difference of all zero. This setting is used in the
experimental setting which the pretrain without reference and the
finetune with difference.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>