Add lora setting in Finetune
If the setting is finetune, it will directly use lora no matter the
sport type is Skating or Boxing. Add a conditional statement in the
dataset to differentiate between skating and boxing. If the sport type
is skating, the dataset should use the revised labels corresponding to
the error segments. If it is boxing, the dataset should use the
original labels.

Reference : https://github.com/microsoft/LoRA