Adjust cross-entropy for ground truth selection
Adjust the method of training in main.py. The logic for handling
cross-entropy loss is adjusted to select the most relevant ground truth
based on logits. From multiple ground truths, the one most similar to
the prediction is selected, and the cross-entropy loss is computed. The
goal is to converge towards the most similar ground truth.

The data loading mechanism in the dataloader was also adjusted to
accommodate the loading of multiple labels, enabling support for
multiple ground truths.

The LoRA parameters for boxing have been adjusted as follows:
Bias: 'none', r: 64 , LoRA alpha: 128, LoRA dropout: 0.5
These adjustments align with the parameters set for both stagcn
and transformation.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>