Fix LoRA finetune not reproducting results
When adapting a pre-trained model to a specific motion task using LoRA
for motion instruction task, the approach involves adding LoRA to all
2D convolution layer in the Human Pose Perception module. This includes
adapting the temporal convolution block in the Human Pose Perception,
where LoRA is applied. However, due to the batch normalization layers on
both sides of the LoRA convolution layer in temporal convolution block,
the embedding forward through this temporal convolution block resulted
in infinite (inf) values, which in turn caused NaN values in the
subsequent third, fourth, and fifth blocks of the perception branch.
This caused an issue where the evaluation could not reproduce the results
from the training process. To resolve this, the temporal convolution
block was reverted to its original structure, using nn.Conv2d, as in the
pre-trained model.

In T5.py, during the generation process, features are forwarded to Human
Pose Perception before checking the PRERTRAIN_DIFFERENCE condition.
Therefore, the mechanism to disable gradient computation and the
mechanism to set the model to evaluation mode were moved earlier in the
code.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>