Update the way of reading usr_start and std_start
Remove the use of the 'standard_longer' variable for determining where
to add 'start_frame'. If the number of frames in the standard video is
greater than the number of frames in the user's video, add 'start_frame'
to 'std_start'. Otherwise, add 'start_frame' to 'usr_start'.

As the dataset format changes, directly use the recorded values in the
dataset: 'std_start_frame', 'start_frame', and 'aligned_seq_len'.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>