Implement if check for Most Similar vs Per-GT Loss
Implement an if statement to choose when the training process will use
the Most Similar ground truth to calculate the loss, or backpropagate
the loss for each ground truth.

The if condition checks the LOSS variable from the config to select
between the first method, denoted by "ClosestSimGT", or the second
method, denoted by "PerGT".

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>