Add six indicators for sport utility evaluation
Introuduce a set of evaluation indicators to assess the sport utility of
the generated instructions. There are six indicators "Error Detection",
"Time Detection", "Body Part Detection", "Causation", "Method" and
"Coordination".

In every evaluation file, we evaluate G-eval score on every instruction
and record their G-eval score in according json file.

In statics.py, we calculate the number of guidance prompts in each
indicator's G-eval and calculate the relationship between two
indicators, counting it as one occurrence when both indicators appear
together in the same instruction, which are presented as an association
matrix.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>