Add debug method for ground truth visualization
When choosing the 'PerGT' strategy to calculate the loss for model
convergence during training, each sample has both an original label
(used for learning) and a best label.

The best label refers to the ground truth that is most similar to the
model's prediction, based on the computed loss. For visualization and
debugging purposes, both the best label and the original label are
saved in corresponding JSON files.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>