Add implementation to handle the evaluation case
Use 'cfg.EVAL.score' to determine whether the current stage is
evaluation. If 'cfg.EVAL.score' is set to 'False', the model is in the
evaluation stage.

In the evaluation stage, there are no target labels in the dataset.
Therefore, it is unnecessary to calculate the CrossEntropyLoss between
the target instruction and the generated instruction.

Since the generate() function in the CoachMe class does not use the
'labels' variable , there is no need to include 'labels' in the 'input'
dictionary, which is the parameter passed to the generate() function.

Signed-off-by: Wei-Hsin Yeh <weihsinyeh168@gmail.com>